@@ -18,84 +18,88 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies: 
   flutter: 
    sdk: flutter 
-  flutter_speed_dial: ^3.0.5
-  google_fonts: ^2.1.0
-  provider: ^5.0.0
-  firebase_core: ^1.3.0
-  cloud_firestore: ^2.2.2
-  firebase_storage: ^8.1.3
-  firebase_auth: ^1.4.1
+  flutter_speed_dial: ^7.0.0
+  google_fonts: ^6.1.0
+  provider: ^6.0.5
+  firebase_core: ^3.10.0
+  cloud_firestore: ^5.6.1
+  firebase_storage: ^12.4.0
+  firebase_storage_web: ^3.6.17
+  firebase_auth: ^5.4.0
   flutter_spinkit: ^5.0.0
-  intl: ^0.17.0
+  intl: ^0.19.0
 
   # 로그인
-  google_sign_in: ^5.0.4
-  pinput: ^1.2.0
-  country_code_picker: ^2.0.2
+  google_sign_in: ^6.2.2
+  pinput: ^5.0.0
+  country_code_picker: ^3.1.0
 
   # GPS
-  geolocator: ^7.2.0+1
+  geolocator: ^13.0.2
 
   # 사진
-  pattern_formatter: ^2.0.0
-  wechat_assets_picker: ^6.0.0-dev.7
-  wechat_camera_picker: ^2.3.0-dev.2
-  image_editor: ^1.6.0
-  permission_handler: ^8.3.0
+  pattern_formatter: ^4.0.0
+  wechat_assets_picker: ^9.4.2
+  wechat_camera_picker: ^4.1.2
+  image_editor: ^1.0.1
+  permission_handler: ^11.3.1
 
   # 이미지 사이즈를 압축해준다.
   flutter_image_compress: ^2.4.0
   # 임시 저장공간을 만들어 준다.(캐시 공간)
   path_provider: ^2.0.2
   # 이미지를 캐시해준다.
-  cached_network_image: ^3.1.0
+  cached_network_image: ^3.3.1
   # pull down시 리스트뷰를 refresh해준다.
   pull_to_refresh: ^2.0.0
   # 무한 스크롤뷰
-  infinite_scroll_pagination: ^3.1.0
+  infinite_scroll_pagination: ^4.0.0
   # 이미지뷰 스크롤
-  carousel_slider: ^4.0.0
+  carousel_slider: ^5.0.0
   # GetX 상태관리 라이브러리
-  get: ^4.3.5
+  get: ^4.6.6
   numberpicker: ^2.1.1
   # 웹뷰
-  webview_flutter: ^2.0.14
+  webview_flutter: ^4.10.0
   # http 통신
-  http: ^0.13.3
+  http: ^1.2.2
   # 디바이스에서 다른 앱을 실행
   url_launcher: ^6.0.9
-  flutter_datetime_picker_bdaya: ^2.0.0
+  flutter_datetime_picker_bdaya: ^3.0.2
   # 텍스트 필드에 마스크를 씌움
   mask_text_input_formatter: ^2.0.0
   # validator
-  string_validator: ^0.3.0
+  string_validator: ^1.1.0
   # 텍스트 접고 펼치기
   expandable_text: ^2.2.0
   # 다이얼로그 창에서 리스트 선택
   select_dialog: ^2.0.0
   # 디바이스에 맞춰 다이얼로그를 다르게 출력
-  adaptive_dialog: ^1.1.0
+  adaptive_dialog: ^2.0.0
   # 채팅을 위한 소켓 통신
-  socket_io_client: ^1.0.1
+  socket_io_client: ^3.0.2
   # 채팅 UI
   flutter_chat_ui: ^1.4.4
   # splash
-  flutter_native_splash: ^1.3.1
-  camera: ^0.9.8+1
-  camera_android: ^0.9.8+3
-
-
-
+  flutter_native_splash: ^2.4.4
+  webview_flutter_android: ^4.2.0
+    # 로그인 안드로이드
+  google_sign_in_android: ^6.1.20
+  
+  sqflite: ^2.3.0
+  
+  video_player: ^2.8.1
+  video_player_android: ^2.4.10  # 최신 버전으로 업데이트
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
 cupertino_icons: ^1.0.2 
-
 
 dev_dependencies:
   flutter_test:
@@ -145,5 +149,3 @@ flutter:
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
-
-
