@@ -1,5 +1,5 @@
 class ApiConfig {
-  static String apiUrl =
-      "http://ec2-13-125-118-200.ap-northeast-2.compute.amazonaws.com:3000/";
-  //static String apiUrl = "http://127.0.0.1:3000/";
+  // static String apiUrl =
+      // "http://ec2-13-125-118-200.ap-northeast-2.compute.amazonaws.com:3000/";
+  static String apiUrl = "http://127.0.0.1:3000/";
 }
